+++ NEWS.md
@@ -3,7 +3,9 @@
 2023-xx-xx
 
 - Fixes/Improvements:
-
+  - GEOSIntersection: Fix FE_INVALID exception on intersection of disjoint geometries
+    (GH-791, Joris Van den Bossche & Dan Baston)
+  - Fix incorrect result from Envelope::disjoint (GH-791, Dan Baston)
 
 ## Changes in 3.11.1
 2022-11-13